+++ extensions/new-file-here/src/new-file-here.tsx
@@ -1,43 +1,159 @@
 import fs from "fs";
 import * as XLSX from "xlsx";
-import { Action, ActionPanel, Icon, List, open, showToast, Toast, useNavigation } from "@raycast/api";
-import { useEffect } from "react";
-import { checkFileExists, getFinderPath, preferences } from "./utils";
-import { codeFileTypes, documentFileTypes, FileType, scriptFileTypes } from "./file-type";
+import { Action, ActionPanel, environment, Icon, List, open, showToast, Toast, useNavigation } from "@raycast/api";
+import React, { useEffect, useState } from "react";
+import { checkDirectoryExists, getFileInfo, getFinderPath, preferences } from "./utils";
+import { codeFileTypes, documentFileTypes, FileType, scriptFileTypes, TemplateType } from "./file-type";
 import { runAppleScript } from "run-applescript";
 import NewFileWithName from "./new-file-with-name";
+import AddFileTemplate from "./add-file-template";
 
 export default function main() {
   const preference = preferences();
+  const templateFolderPath = environment.supportPath + "/templates";
+  const [templateFiles, setTemplateFiles] = useState<TemplateType[]>([]);
+  const [updateList, setUpdateList] = useState<number[]>([0]);
+  const [isLoading, setIsLoading] = useState<boolean>(true);
+  const { push } = useNavigation();
 
   useEffect(() => {
     async function _initRunAppleScript() {
+      const _templateFiles: TemplateType[] = [];
+      try {
+        if (await checkDirectoryExists(templateFolderPath)) {
+          fs.readdirSync(templateFolderPath).forEach((file) => {
+            if (!file.startsWith(".")) {
+              const filePath = templateFolderPath + "/" + file;
+              const { nameWithoutExtension, extension } = getFileInfo(filePath);
+              _templateFiles.push({
+                path: filePath,
+                name: nameWithoutExtension,
+                extension: extension,
+                simpleContent: false,
+              });
+            }
+          });
+          setIsLoading(false);
+        } else {
+          fs.mkdirSync(templateFolderPath);
+        }
+      } catch (e) {
+        await showToast(Toast.Style.Failure, String(e));
+      }
+      setTemplateFiles(_templateFiles);
       await runAppleScript("");
     }
 
     _initRunAppleScript().then();
-  }, []);
+  }, [updateList]);
 
   return (
-    <List isShowingDetail={false} isLoading={false} searchBarPlaceholder={"Search File Type"}>
+    <List
+      isShowingDetail={false}
+      isLoading={isLoading}
+      searchBarPlaceholder={"Search File"}
+      selectedItemId={templateFiles.length > 0 ? templateFiles[0].path : ""}
+    >
+      <List.Section title={"Template"}>
+        {templateFiles.map((template, index, array) => {
+          return (
+            <List.Item
+              id={template.path}
+              key={template.path}
+              icon={{ fileIcon: template.path }}
+              title={template.name}
+              subtitle={template.extension}
+              actions={
+                <ActionPanel>
+                  <Action
+                    title={"New File Here"}
+                    icon={Icon.Finder}
+                    onAction={async () => {
+                      try {
+                        await createNewFileByTemplate(template, await getFinderPath());
+                      } catch (e) {
+                        await showToast(Toast.Style.Failure, "Create file failure.", String(e));
+                      }
+                    }}
+                  />
+                  <Action
+                    title={"New File with Name"}
+                    icon={Icon.Document}
+                    onAction={() => {
+                      push(
+                        <NewFileWithName newFileType={{ section: "Template", index: index }} templateFiles={array} />
+                      );
+                    }}
+                  />
+                  <Action
+                    title={"Add File Template"}
+                    icon={Icon.TextDocument}
+                    shortcut={{ modifiers: ["cmd"], key: "n" }}
+                    onAction={() => {
+                      push(<AddFileTemplate updateListUseState={[updateList, setUpdateList]} />);
+                    }}
+                  />
+                  <Action
+                    title={"Delete File Template"}
+                    icon={Icon.Trash}
+                    shortcut={{ modifiers: ["shift", "cmd"], key: "backspace" }}
+                    onAction={async () => {
+                      await showToast(Toast.Style.Animated, "Deleting template...");
+                      fs.unlinkSync(template.path);
+                      const _templateFiles = [...templateFiles];
+                      _templateFiles.splice(index, 1);
+                      setTemplateFiles(_templateFiles);
+                      await showToast(Toast.Style.Success, "Delete template success!");
+                    }}
+                  />
+                </ActionPanel>
+              }
+            />
+          );
+        })}
+      </List.Section>
       {preference.showDocument && (
         <List.Section title={"Document"}>
-          {documentFileTypes.map((fileType) => {
-            return <FileTypeItem key={fileType.languageId} fileType={fileType} />;
+          {documentFileTypes.map((fileType, index) => {
+            return (
+              <FileTypeItem
+                key={fileType.languageId}
+                fileType={fileType}
+                newFileType={{ section: "Document", index: index }}
+                templateFiles={templateFiles}
+                updateListUseState={[updateList, setUpdateList]}
+              />
+            );
           })}
         </List.Section>
       )}
       {preference.showCode && (
         <List.Section title={"Code"}>
-          {codeFileTypes.map((fileType) => {
-            return <FileTypeItem key={fileType.languageId} fileType={fileType} />;
+          {codeFileTypes.map((fileType, index) => {
+            return (
+              <FileTypeItem
+                key={fileType.languageId}
+                fileType={fileType}
+                newFileType={{ section: "Code", index: index }}
+                templateFiles={templateFiles}
+                updateListUseState={[updateList, setUpdateList]}
+              />
+            );
           })}
         </List.Section>
       )}
       {preference.showScript && (
         <List.Section title={"Script"}>
-          {scriptFileTypes.map((fileType) => {
-            return <FileTypeItem key={fileType.languageId} fileType={fileType} />;
+          {scriptFileTypes.map((fileType, index) => {
+            return (
+              <FileTypeItem
+                key={fileType.languageId}
+                fileType={fileType}
+                newFileType={{ section: "Script", index: index }}
+                templateFiles={templateFiles}
+                updateListUseState={[updateList, setUpdateList]}
+              />
+            );
           })}
         </List.Section>
       )}
@@ -45,9 +161,17 @@ export default function main() {
   );
 }
 
-function FileTypeItem(props: { fileType: FileType }) {
+function FileTypeItem(props: {
+  fileType: FileType;
+  newFileType: { section: string; index: number };
+  templateFiles: TemplateType[];
+  updateListUseState: [number[], React.Dispatch<React.SetStateAction<number[]>>];
+}) {
   const { push } = useNavigation();
   const fileType = props.fileType;
+  const newFileType = props.newFileType;
+  const templateFiles = props.templateFiles;
+  const [updateList, setUpdateList] = props.updateListUseState;
   return (
     <List.Item
       icon={{ source: fileType.icon }}
@@ -61,7 +185,7 @@ function FileTypeItem(props: { fileType: FileType }) {
               try {
                 await createNewFile(fileType, await getFinderPath());
               } catch (e) {
-                await showToast(Toast.Style.Failure, "Create File Failed", String(e));
+                await showToast(Toast.Style.Failure, "Create file failure.", String(e));
               }
             }}
           />
@@ -69,7 +193,15 @@ function FileTypeItem(props: { fileType: FileType }) {
             title={"New File with Name"}
             icon={Icon.Document}
             onAction={() => {
-              push(<NewFileWithName fileType={fileType} />);
+              push(<NewFileWithName newFileType={newFileType} templateFiles={templateFiles} />);
+            }}
+          />
+          <Action
+            title={"Add File Template"}
+            icon={Icon.TextDocument}
+            shortcut={{ modifiers: ["cmd"], key: "n" }}
+            onAction={() => {
+              push(<AddFileTemplate updateListUseState={[updateList, setUpdateList]} />);
             }}
           />
         </ActionPanel>
@@ -78,7 +210,7 @@ function FileTypeItem(props: { fileType: FileType }) {
   );
 }
 
-export function createFileName(fileType: FileType) {
+export function createFileName(name: string, extension: string): string {
   const date = new Date();
   const time =
     date.getFullYear() +
@@ -92,17 +224,18 @@ export function createFileName(fileType: FileType) {
     date.getMinutes() +
     "" +
     date.getSeconds();
-  return fileType.name + "-" + time + fileType.extension;
+  return name + "-" + time + "." + extension;
 }
 
 export async function createNewFile(
   fileType: FileType,
   path: string,
-  fileName = createFileName(fileType),
+  fileName = createFileName(fileType.name, fileType.extension),
   fileContent = ""
 ) {
+  await showToast(Toast.Style.Animated, "Creating file...");
   const filePath = path + fileName;
-  const isExist = await checkFileExists(filePath);
+  const isExist = await checkDirectoryExists(filePath);
   if (!isExist) {
     if (fileType.name === "Excel") {
       const workbook = XLSX.utils.book_new();
@@ -113,25 +246,42 @@ export async function createNewFile(
       fs.writeFileSync(filePath, fileContent);
     }
   }
+  await showCreateToast(isExist, filePath, path);
+}
 
+export async function createNewFileByTemplate(
+  template: TemplateType,
+  folderPath: string,
+  srcPath = template.path,
+  fileName = createFileName(template.name, template.extension)
+) {
+  await showToast(Toast.Style.Animated, "Creating file...");
+  const filePath = folderPath + fileName;
+  const isExist = await checkDirectoryExists(filePath);
+  if (!isExist) {
+    fs.copyFileSync(srcPath, filePath);
+  }
+  await showCreateToast(isExist, filePath, folderPath);
+}
+const showCreateToast = async (isExist: boolean, filePath: string, folderPath: string) => {
   const options: Toast.Options = {
     style: isExist ? Toast.Style.Failure : Toast.Style.Success,
-    title: isExist ? "File already Exists" : "Create File Success",
+    title: isExist ? "File already exists." : "Create file success!",
     message: "Click to open file",
     primaryAction: {
-      title: "Open File",
-      onAction: (toast) => {
-        open(filePath);
-        toast.hide();
+      title: "Open file",
+      onAction: async (toast) => {
+        await open(filePath);
+        await toast.hide();
       },
     },
     secondaryAction: {
-      title: "Reveal in Finder",
-      onAction: (toast) => {
-        open(path);
-        toast.hide();
+      title: "Reveal in finder",
+      onAction: async (toast) => {
+        await open(folderPath);
+        await toast.hide();
       },
     },
   };
   await showToast(options);
-}
+};