+++ extensions/new-file-here/CHANGELOG.md
@@ -1,5 +1,7 @@
-# New File Here Changelog
+# Easy New File Changelog
 
-## [Added New File Here] - 2022-04-01
+## [Added Easy New File] - 2022-04-10
 - Release of New File Here version 1.0
 - **New File Here**: Create a new file in the current Finder window or on the desktop.
+  - New File with Name: Create a new file with name.
+- **Add File Template**: Add a file template that you can use in the "New File Here" command.